test: consolidate encoded_len* tests
Move the encoded_len_limit tests to the tests-overflow crate, as well as
the encoded_len.proto file used to generate the testbed.

I was getting carried away with testing these limits, up to adding Kani
proofs, but the added CI complexity is not worth it. It could have been
a boolean encoded_len_limited method for what matters, as the number
of fields in the message is limited enough so that overflowing unchecked
arithmetic on lengths of the numeric fields should never be a concern
in practice.

Still, the tests-overflow crate is close enough thematically
and is not saddled with a dependency on protobuf which takes a long
time to build.
