+++ tests/overflow/src/encoded_len.proto
